+++ src/App.tsx
@@ -3,14 +3,8 @@ import '@fortawesome/fontawesome-free/css/all.css';
 import './App.scss';
 import { Navigate, Route, Routes } from 'react-router-dom';
 import { HomePage } from './components/HomePage';
-import { TabsPage } from './components/TabsPages';
+import { TabsPage } from './components/TabsPage';
 import { Navbar } from './components/Navbar';
-
-const tabs = [
-  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
-  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
-  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
-];
 
 export const App = () => (
   <>
@@ -19,15 +13,15 @@ export const App = () => (
     <div className="section">
       <div className="container">
         <Routes>
-          <Route path="*" element={<h1 className="title">Page not found</h1>} />
-
           <Route path="/" element={<HomePage />} />
           <Route path="home" element={<Navigate to="/" replace />} />
 
           <Route path="tabs">
-            <Route index element={<TabsPage tabs={tabs} />} />
-            <Route path=":activeTabId?" element={<TabsPage tabs={tabs} />} />
+            <Route index element={<TabsPage />} />
+            <Route path=":activeTabId?" element={<TabsPage />} />
           </Route>
+
+          <Route path="*" element={<h1 className="title">Page not found</h1>} />
         </Routes>
       </div>
     </div>
