--- src/components/TabsPages/index.ts
@@ -1,1 +1,0 @@
-export * from './TabsPages';
